pyproject.toml
[tool.poetry]
name = "9gagbot"
version = "0.1.0"
description = ""
authors = ["Danilo Shiga <daniloshiga@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.4"
aiofiles = "^0.4.0"
belogging = "^0.1.3"
aiogram = "^2.0"
tortoise-orm = "^0.12.1"
aiosqlite = "^0.10.0"
prettyconf = "^2.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.11.0"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parenthesis = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
